post

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,23 @@ class TasksController extends Controller
      **/
     public function index()
     {
-        $tasks = Task::all();
-        
-         return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+         if (\Auth::check()) {
+$user = \Auth::user();
+           
+$tasks = $user->tasks()->get();
+$data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', $data);
+    } 
+     else{
+           return view('welcome');
+    }
     }
     
+    
+   
     
     /**
      * Show the form for creating a new resource.
@@ -51,11 +61,11 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         $task = new Task;
+        $task->user_id = \Auth::id();
         $task->status = $request->status; 
         $task->content = $request->content;
         $task->save();
-
-    
+       
         return redirect('/');
     }
 
@@ -68,10 +78,14 @@ class TasksController extends Controller
     public function show($id)
     {
          $task = Task::findOrFail($id);
-
-        return view('tasks.show', [
+         if (\Auth::id() === $task->user_id) {
+             return view('tasks.show', [
              'task' => $task,
              ]);
+        }
+        else {
+            return redirect('/');
+          }
     }
     /**
      * Show the form for editing the specified resource.
@@ -81,12 +95,16 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-         $task = task::findOrFail($id);
-
-        // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
+       $task = Task::findOrFail($id); 
+if (\Auth::id() === $task->user_id) {
+    return view('tasks.edit', [
             'task' => $task,
         ]);
+}
+else {
+     return redirect('/');
+}
+       
     }
 
     /**
@@ -103,7 +121,7 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
 
-        $task = task::findOrFail($id);
+        $task = Task::findOrFail($id); 
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -120,9 +138,10 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-         $task = Task::findOrFail($id);
-        
-        $task->delete();
+         $task = Task::findOrFail($id); 
+         if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
     
         return redirect('/');
